Add inertia tensors to AirfoilGeometry
This is intended to support the computation of Parafoil inertia tensors.
If you treat each section of the parafoil as an Airfoil that has been
stretched on an axis normal to its plane, then you merely need to scale
the airfoil's inertia and adjust Ixx and Izz for the parafoil section.

This also introduces upper and lower curve lengths for the airfoils,
which can also be used to compute the upper and lower surface area of
the 3d wing.