Major refactor: reparametrize ParafoilGeometry
Instead of defining the parafoil geometries in terms of the y-coordinate
of the inflated wing, convert ellipses to their parametric forms, and
update all functions in terms of the new parametric parameter `t`.

For simplicity, let `t` be defined between -1 and 1, corresponding to
the left and right wing tips, so `t = 0` corresponds to the wing root.
This requires careful conversion between parameter forms (since the
ellipse equations are defined in terms of a `t` that varies between 0 to
pi, so the function parameter `t` and the ellipse parameter `t` are not
the same), but I think it's worth the effort.