WIP: refactor and reparametrize ParafoilGeometry
This is a broken build as a development checkpoint.

This is a ground up rework that is separating the ParafoilGeometry
into Planform and Lobe components. It is also reparametrizing the
entire API in terms of normalized flattened span coordinates, `s`,
where `s = 1` is the far right wing tip, and `s = -1` is the far left.

This reparametrization should simplify the definition and implementation
of Lobe deformations (such as for weight shift), since the spanwise
coordinates are now constant (for example, `s = 0.63` is always 63% of
the way between the wing root and the wing tip).