Replace geometric torsion distribution
I was choosing between 'linear' and 'exponential', but linear isn't very
useful, and exponential produced a derivative discontinuity at the wing
root, so I've replaced it with a simple `torsion_max * x^n` model. It's
not very realistic for construction purposes, but it gives good control
over the development of the torsion at the wing tips.