Add Cubic BrakeGeometry and switch to delta angles
First, the Cubic geometry replaces all the other rotten versions; it's
reasonably configurable for decent-looking deflection magnitudes.

Second, I cleaned up the API by making it explicit that the
BrakeGeometry returns the final delta angles for each chord, even though
it doesn't actually know what the chord *is*. This does not work with
physical line geometries, which can only deal with deflection distances,
but it's a reasonable stand-in for now.