Remove ParafoilSections
It's much too complicated to create a smart container that generates the
airfoils and forwards their API in an intelligent way.

What's more, it's very unlikely that I'll get around to creating a
generation method for the airfoil coefficients and geometries. It just
seems much too unlikely to warrant inclusion in the current design, so
I'll try to let the Airfoil methods handle the broadcasting instead.