Fix the `MAC_to_c0` equation
I was using the equation from Paraglider Flight Dynamics, but it is
giving me the wrong answers. I'm not certain I don't have a bug
somewhere, but when I use my own derivation for this function I get the
correct planform surface area. For now I've left the old code.