Add inertia calculations for Parafoil and Wing
I have equations from Paraglider Flight Dynamics to estimate the inertia
tensor of the wing, but there was no justification, and I suspected it
wouldn't work well for the products of inertia (the Ixz terms). I went
ahead and wrote an (exhausting) method to compute the individual
inertias for the two surfaces and the internal volume.

It currently shows a smaller Ixz than the PFD method (by a factor of
about 2.5, typically), but that is only a relative measure, since I
don't know for certain what "wing density" to use. I'm sure I'm not
accounting for all the mass, so it's very likely I'm underestimating the
inertias.

Not sure which method to use right now; my new method is a beast.