Reorganize the Airfoil buildup
Since I'm using the AirfoilGeometry for computing the inertia, I should
make it more clear that choosing the geometry is not arbitrary. I
reorganized the pieces to group coefficients+geometry more closely.

I also added an unfiltered set of coefficients for the NACA4415 and
NACA2412. (Like the 4412, they are pretty messy at the extreme ranges of
alpha, due to convergence issues in XFOIL.)