Fix the speedbar geometry
The speedbar was using the leading and trailing edge as the line
connection points, but they should be where the A and C lines attach to
the central chord. Using the LE and TE means the accelerator would need
to be very long to produce the necessary change in pitch.

This way, the wing rotates around the C connection.