Reduce the AirfoilGeometry API
I'm redesigning the AirfoilGeometry to rely on a set of tabulated points
instead of equations (although you can generate the points using
equations, such as for NACA foils). There are subtleties regarding the
definition of the chord, camber line, and thickness, so for now I'm
removing them from the API instead of trying to guess.

For now, any users that require the thickness or camber line will need
to do it themselves.