Add a 3D "braking wing" plot
The standard "plot a wing" function doesn't understand how to modify the
wing sections when the brakes are being applied. This hack was a
starting point to show how it might be done.

This has an embedded "curving airfoil" code that I'm almost definitely
not going to be using, but I'm keeping this script for now in case it's
useful earlier.