Fix cambered airfoils with max camber at the nose
The second digit, P, specifies where the maximum camber occurs. The
camber lines are defined in terms of `p`. If `p = 0`, the equation for
points forward (towards the leading edge) of that point `p` is
undefined, but since those would be invalid `x` coordinates anyway, we
simply don't evaluate that equation.