WIP: swapping in new ParafoilGeometry
This new version generalizes the geometry to use (almost) arbitrary
functions for the chord, planform, and lobe. This makes it easier to use
explicit point-definitions to describe a wing instead of trying to find
a best-fit Elliptical version. (For example, in Belloc he provides a set
of points directly on the 0.6c curve; this new geometry lets you use
those directly, through an interpolator.)

It does still require that the lobe curve also provides `dy/ds` and
`dz/ds`, but those are simple enough with elliptical functions. See
`belloc.py` for an example that uses a `PchipInterpolator` via a trivial
interface.