Enable raw airfoil coordinates in `surface_points`
`FoilGeometry.surface_points` was originally designed to transform input
coordinates to either the airfoil coordinates of either the upper or
lower surfaces, but that requires two calls to get to both sides of the
foil. Sometimes the user may want a continuous curve, such as when
generating a 3D mesh.