Rename "thickness" function for NACA airfoils
The NACA definitions are a pain.

1. They have two different conventions for measuring thickness

2. The "American" convention make finding the camber curve non-trivial.

Instead of letting the NACA convention dictate the behavior of the
entire program, isolate the "thickness" definition as a concept internal
to the NACA geometry itself. I'm using `_yt` since that's the naming
most common for NACA definitions, but `_yt` no longer overrrides the
global `AirfoilGeometry::thickness` function.