Iterate to account for varying Reynolds numbers
The Paraglider equation was originally developed without consideration
for Reynolds numbers. Once they are accounted for, you can't just
compute the forces using `V == 1` and scale, since the equilibrium
depends on both the angle and the airspeed.

The original equation is pretty close though, so just a few iterations
are plenty to account for the change.