WIP: replacing `Paraglider.forces_and_moments`
The individual components (wing, harness) will return forces and
moments, but the `Paraglider` will focus on accelerations. I'm working
towards a generalized 9-DoF model, and it won't make sense to put the
dynamics outside the class (as currently happens with the poorly named
`GliderSim` class).

The equilibrium also needs to be overhauled: it doesn't account for the
weight of the wing (the old method solved for `alpha_eq` independent of
pitch angle or airspeed, but you really need to solve for all three
simultaneously) and doesn't allow for relative orientations of the
harness.