Add a new `equilibrium_glide` method to Paraglider
The other method was based on Benedetti, but ignored the weight of the
wing. In order to account for the weight of the wing, you need a
non-zero pitching moment and a specific airspeed, but you those values
also depend on `theta`, since that changes the gravity vector.

This new method is pretty ugly, and SLOW, but for now it works.