Overhaul `Paraglider.accelerations`
Major cleanup in preparation for a 9-DoF model based on "Evaluation of
Multibody Parafoil Dynamics Using Distributed Miniature Wireless
Sensors" (Slegers and Costello, 2012). The idea will be to build a
system of linear equations that connect the translational and angular
accelerations to the net forces and moments, which will enable splitting
the canopy from the payload and treating them as two independent, but
connected, rigid bodies. I mostly just want the relative pitch, but I
don't see any reason not to just allow for the full 9-DoF model.