Standardize some vector naming
Things like `UVW` and `PQR` fail to map onto the equations well, and
don't communicate their real definition, so I'm replacing them with
`v_cm2e` and `omega_b2e`. This will become even more important once the
reference velocity moves away from the cm to the risers (so the
traditional `UVW` naming would be misleading).

I'm also using `alpha_b2e` instead of `omega_dot` since it matches what
you're likely to write in an equation. There is a bit of overlap with
`alpha` for angle of attack, but in practice it's not a big deal:
vectors have the `a2b` subscripts, scalars either have no subscript or a
simple label (no "a with respect to b" notation).