Encode payload orientation in absolute terms
Using `q_p2b` is a little convenient when you're computing `Theta_p`,
but in all other cases it's unnecessarily complicated. Using `q_p2e`
maintains symmetry with `q_b2e` which makes the code easier to follow,
plus it's easier to compute positions on the payload relative to Earth.