Fix payload translational acceleration
I mistakenly took the derivative of `v_P2e` with respect to the payload,
but you have to use the same frame of reference for the derivatives of
both `v_B2e` and `v_P2e`, since that's what the linear system is solving
for. Slegers chose to take them both with respect to the body, so that's
what I've copied here.