Revert to using `q_p2b` for the state
Turns out that trying to compute the physically-relative angles from the
two global angles `q_b2e` and `q_p2e` is a pain (just computing their
difference and converting it to a set of yaw-pitch-roll Euler angles
does produce a rotation that would invert the rotation, but it's not the
physically relevant angles). I'm sure you it can be done, but I'm in a
hurry.