Overhaul the equilibrium code
The primary change is to return the equilibrium conditions as a
dictionary; using individual variables for each of the factors was
getting messy, especially when you'd like two different sets of
equilibrium values for code that have references to the two different
paraglider models.

It also gave me the chance to cleanup `plot_polar_curve` a bit, which
was a monstrosity from day one.