Add control points for line drag
These are a bit crude, but they stabilize turning with weight shift and
they bring the polar curves more in line with what I'd expect
(considering this is a naive approximation of the real wing, not the
true, optimized design).

The line drag coefficient is taken from "Identification of a degradation
of aerodynamic characteristics of a paraglider due to its flexibility
from flight test" (Kulhánek, 2019).