Add more paraglider models
`Paraglider6b` uses the angular momentum of the system about the origin
(the riser connection `R`) instead of the body center of mass, which
makes it easier incorporate apparent inertia effects.

`Paraglider6c` computes the angular momentum about the body center of
mass `B` and solves for `a_B2e` and `alpha_b2e` (before computing `a_R2e`
from those) which enables fully decoupled linear and angular momentum
equations (which are more common/straightforward, making this a good way
to test the other more complicated models).

`Paraglider9b` is similar to `Paraglider6b`: it computes the angular
momentums for both the body and the payload about the riser connection
`R`, which will make it easier to incorporate apparent inertia for the
wing.