Use `r` as standard airfoil curve parameter
This is a general purpose parameter for points on the chords, camber
lines, and surfaces. Each curve is a parametric curve of `r`, the
normalized curve length. `0` is always the leading edge, `1` is the
trailing edge.

In the special case of "airfoil" coordinates, `1` is the upper surface
TE, and `-1` is the lower surface TE.