Override `__dir__` in modules
I'm tired of tab-completion listing every identifier in module scope, so
I'm using PEP 562 (Python 3.7) to override `__dir__` to return just the
explicitly exported symbols as defined in `__all__`.

I don't love this solution since you have to keep `__all__` synchronized
with the actual symbols, but `dir()` is mostly for user-convenience
anyway so it should only be irritating at worst.