Eliminate `airfoil.Airfoil`
It was a useless wrapper class. The idea behind it was to bundle a
single airfoil (the shape) with the coefficients for that shape. The
problem with that was that it's slow and awkward to manually interpolate
the coefficients for different values of `delta_f`. It's much easier and
faster to use a single gridded table of coefficients with a
`RegularGridInterpolator`, but now you have a mismatch: the coefficients
support `delta_f`, the geometries do not.

I decided to focus on the primary goal of the project: flying gliders.
Using flappless airfoil coefficients should be the exception rather than
the rule.

There's still the issue that the user must provide profiles and
coefficients that agree, but I don't see that it makes much difference
if that happens when instantiating an `Airfoil` or a `FoilSections`.