Replace crude `equilibrium_state` implementations
The old methods were clumsy evolutions that arose from old model design
issues. They're inaccurate (and slow, in the case of the simulator-based
methods). Besides, the simulator methods duplicate code in `simulator`
anyway.

The old `Paraglider6a.equilibrium_state2` might be interesting in some
context since it's a closed form solution, but I don't like that it
doesn't use the actual model dynamics, so Paraglider6b et al would have
no effect on the equilibrium state (which is currently valid, but in the
future the dynamics may affect the equilibrium state; better to err on
the side of explicit).