Use controls at `t = 0` for computing equilibrium
It's a lot easier to start the wing in a nice state than to applies some
controls, wait for the system to settle, and only then execute the
desired input sequence.

For example, a rapid transition from full-brake to zero-brake would
require applying full-brake, waiting some fixed delay for equilibrium to
hopefully exist, and then remove the brakes. This change lets you
initialize the scenario in the full-brake equilibrium.