Tweak Hook3 specification
The tweaked `r_x` is from overlaying the edge outline onto the flattened
image from the manual; hard to say if it is more accurate, but it's a
minor change.

Added a potentially (slightly) more accurate `yz` curve that increases
`tip_anhedral` (after trying to tweak the parameters to make the
projected area and projected span match the specs), but the more extreme
angles at the wing tips make it more likely to stall without providing
any significant benefit to the overall model accuracy. For now I'm
leaving that disabled and sticking with a purely circular arc.