Separate EllipticalArc and EllipticalChord
The original, abstracted `EllipticalArc` (where "arc" in that case meant
"elliptical arc segment", not the "arc" of a parafoil) was misguided,
and never ended up being useful. It was confusing, and what's worse is
it was wrong! (It assumed linear spacing in `t` produced linear spacing
in `s`.)

The abstraction was unnecessary, so I've refactored the arc and chord
functors into simple, dedicated classes. Much nicer (and correct).