Redesign the LineGeometry to produce distances
Because the SimpleLineGeometry assumed uniform angles, it was ignoring
the fact that the deflection angle depends on the chord length. The
problem is that the LineGeometry doesn't know the chord lengths; it's
interested in computing deflection distances.

The solution is to compute normalized deflection distances and let the
ParagliderWing determine the resulting deflection angles. It's not
pretty, but it works.