Use Steven's definition of quaternion inverse
In practice, negative the scalar part of the quaternion often has the
same effect as negating the vector part, but in some cases it breaks
equations. For example, Stevens Eq:1.8-21 expects the inverse to be
defined as in Eq:1.8-5 (negating the vector part). If you invert the
scalar part, the magnitudes of the components will be correct but their
signs will be reverse.