Replace delta_f with delta_d
Using the deflection angle `delta_f` was not accurate because it
neglected deflection in the `x` direction. By computing `delta_f` using
the undeflected chord length it was underestimating the angles produced
for a given brake deflection.

Instead, use the vertical deflection distance directly. (Coefficients
expect the distance to be normalized for a unit chord.)