Add alpha "clamping" to the section coefficients
Phillips' method generates fictitious spikes in the induced velocity
which causes spikes in the local angles of attack. The result is a
tendency to artificially stall the wingtips during any maneuver that
isn't slow enough.

The effect can be mitigated by assuming that if the alpha at the wing
tips have exceeded the range supported by the coefficient data, then
"clamp" alpha to its largest value; for Cl/Cd/Cm this returns a constant
value, for Cl_alpha it will produce a zero.

The purpose is not to "fix" the error, the intent is to mitigate it. For
a sufficiently large number of wing segments, the area of the wing tip
should be negligible (especially for a tapered wing), so errors at the
tip should also be negligible. (This is why I chose to specify clamping
using a section index instead of a segment index: it's less likely to
encourage accidental clamping of large wing tip segments.)

It's also important to only clamp the wing tips because the simulator
SHOULD fail when encountering stall conditions. In the even that the
angle of attack is genuine, the non-clamped sections near the wing tip
will still "fail" correctly.