Eliminate `r_CP2RM` from `Paraglider.accelerations`
The original intent was 1) to avoid recomputation, and 2) draw attention
to the fact that distinct wind vectors must be associated with specific
control point positions, but 1) the recomputation cost is negligible,
and 2) if a caller isn't keeping the inputs synced then passing them
probably wouldn't help anyway. The cost just doesn't seem worth it.