Replace ParagliderHarness.aerodynamics with resultant_force
The harness and wing are rigid bodies, so there's not advantage in
passing all the individual forces back to the paraglider model. It just
makes it messy and adds duplicate code to the models.

There's not great name for it, but at least "resultant force" is a
fairly standard term for the net force and moment on a system.